Refactor logger initialization and move it to core module
